Further fixes for instrument copying
