+++ reduction_files/DG_reduction.py
@@ -18,7 +18,7 @@
 from mantid.simpleapi import *
 from mantid.api import AnalysisDataService as ADS
 import numpy as np
-import re, os, h5py
+import os, sys
 import time
 from importlib import reload
 
@@ -69,51 +69,6 @@ def tryload(irun):              # loops till data file appears
         break
 #========================================================
 
-#========================================================
-# Functions to copy instrument info needed by HORACE
-# Resolution convolution if it exists in the raw file
-def get_raw_file_from_ws(ws):
-    for alg in [h for h in ws.getHistory().getAlgorithmHistories() if 'Load' in h.name()]:
-        for prp in [a for a in alg.getProperties() if 'Filename' in a.name()]:
-            if re.search('[0-9]*.nxs', prp.value()) is not None:
-                return prp.value()
-    raise RuntimeError('Could not find raw NeXus file in workspace history')
-
-def copy_inst_info(outfile, in_ws):
-    try:
-        raw_file_name = get_raw_file_from_ws(mtd[in_ws])
-    except RuntimeError:
-        return
-    print(raw_file_name)
-    if not os.path.exists(outfile):
-        outfile = os.path.join(mantid.simpleapi.config['defaultsave.directory'], os.path.basename(outfile))
-    print(outfile)
-    with h5py.File(raw_file_name, 'r') as raw:
-        with h5py.File(outfile, 'r+') as spe:
-            exclude = ['dae', 'detector_1', 'name']
-            to_copy = [k for k in raw['/raw_data_1/instrument'] if not any([x in k for x in exclude])]
-            print(spe.keys())
-            spe_root = list(spe.keys())[0]
-            en0 = spe[f'{spe_root}/instrument/fermi/energy'][()]
-            if 'fermi' in to_copy:
-                del spe[f'{spe_root}/instrument/fermi']
-            for grp in to_copy:
-                print(grp)
-                src = raw[f'/raw_data_1/instrument/{grp}']
-                h5py.Group.copy(src, src, spe[f'{spe_root}/instrument/'])
-            if 'fermi' in to_copy:
-                spe[f'{spe_root}/instrument/fermi/energy'][()] = en0
-            detroot = f'{spe_root}/instrument/detector_elements_1'
-            spe.create_group(detroot)
-            for df0, df1 in zip(['SPEC', 'UDET', 'DELT', 'LEN2', 'CODE', 'TTHE', 'UT01'], \
-                ['spectrum_number', 'detector_number', 'delt', 'distance', 'detector_code', 'polar_angle', 'azimuthal_angle']):
-                src = raw[f'/raw_data_1/isis_vms_compat/{df0}']
-                h5py.Group.copy(src, src, spe[detroot], df1)
-            for nn in range(raw['/raw_data_1/isis_vms_compat/NUSE'][0]):
-                src = raw[f'/raw_data_1/isis_vms_compat/UT{nn+1:02d}']
-                h5py.Group.copy(src, src, spe[detroot], f'user_table{nn+1:02d}')
-#========================================================
-
 # ==============================setup directroties================================
 config['default.instrument'] = inst
 if save_dir is not None:
@@ -143,10 +98,19 @@ else:
 
 cycle_shortform = cycle[2:] if cycle.startswith('20') else cycle
 datadir = f'/archive/NDX{inst}/Instrument/data/cycle_{cycle_shortform}/'
-mapdir  = f'/usr/local/mprogs/InstrumentFiles/{inst.swapcase()}/'
+instfiles  = '/usr/local/mprogs/InstrumentFiles/'
+mapdir  = f'{instfiles}/{inst.swapcase()}/'
 config.appendDataSearchDir(mapdir)
 config.appendDataSearchDir(datadir)
 ws_list = ADS.getObjectNames()
+
+# Try to load subsidiary utilities
+sys.path.append(instfiles)
+try:
+    from reduction_utils import *
+    utils_loaded = True
+except ImportError:
+    utils_loaded = False
 
 print(f'\n======= {inst} data reduction =======')
 print(f'Working directory... {ConfigService.Instance().getString("defaultsave.directory")}\n')
@@ -325,9 +289,10 @@ for irun in sample:
         print(f'{inst}: Writing {ofile}{saveformat}')
         if saveformat.lower() == '.nxspe':
             SaveNXSPE('ws_out', ofile+saveformat, Efixed=Ei, KiOverKfScaling=True)
+            if utils_loaded:
+                copy_inst_info(ofile+saveformat, 'ws_out')   # Copies instrument info for Horace
         elif saveformat.lower() == '.nxs':
             SaveNexus('ws_out', ofile+saveformat)
-            copy_inst_info(ofile+saveformat, 'ws_out')
         if QENS:
             print('... outputting QENS "_red" format')
             theta = np.array([theta_range[0], theta_range[1]])*np.pi/180.
